Moved electron/ui styles into dedicated directory
