+++ electron/src/ui/styles/App.css
